remove associated projects section on resource details if no projects
