+++ modules/research-framework/portal/src/components/repositories/RepositorySpecificDetails.tsx
@@ -1,26 +1,18 @@
-import { RepositoryResource } from "@/interfaces/ResourceType";
-import { Box, Heading, Separator } from "@chakra-ui/react";
-import { AssociatedProjectsSection } from "../projects/AssociatedProejctsSection";
-import { GitHubFileTree } from "./GitHubFileTree";
+import {RepositoryResource} from "@/interfaces/ResourceType";
+import {Box} from "@chakra-ui/react";
+import {AssociatedProjectsSection} from "../projects/AssociatedProjectsSection.tsx";
+import {GitHubFileTree} from "./GitHubFileTree";
 
 export const RepositorySpecificDetails = ({
-  repository,
-}: {
+                                            repository,
+                                          }: {
   repository: RepositoryResource;
 }) => {
   return (
-    <Box>
       <Box>
-        <Heading fontWeight="bold" size="2xl" mb={2}>
-          Associated Projects
-        </Heading>
+        <AssociatedProjectsSection resourceId={repository.id}/>
 
-        <AssociatedProjectsSection resourceId={repository.id} />
+        <GitHubFileTree repository={repository}/>
       </Box>
-
-      <Separator my={6} />
-
-      <GitHubFileTree repository={repository} />
-    </Box>
   );
 };
